# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: DeenMate API
  description: |
    Comprehensive Islamic application backend providing access to Quran, Hadith, Prayer Times, 
    Zakat calculations, and Islamic audio content. This API is designed to serve Islamic 
    applications, websites, and mobile apps with accurate, reliable, and well-structured Islamic data.
  
    ## Features
    - **Quran Module**: Complete Quran text, translations, and recitations
    - **Prayer Module**: Accurate prayer times, qibla direction, and calculation methods
    - **Hadith Module**: Authentic hadith collections and search functionality
    - **Zakat Module**: Zakat calculations and current market values
    - **Audio Module**: High-quality Islamic audio content and metadata
  
    ## Authentication
    Currently operates in read-only mode without authentication. Future versions will include 
    JWT-based authentication for user-specific features.
  
    ## Rate Limiting
    - Public Endpoints: 100 requests per minute per IP
    - Authenticated Endpoints: 1000 requests per minute per user
    - Premium Tier: 5000 requests per minute per user
  
    ## Support
    - Documentation: https://docs.deenmate.com
    - API Status: https://status.deenmate.com
    - Support Email: api-support@deenmate.com
    - Developer Discord: https://discord.gg/deenmate
  
  version: 1.0.0
  contact:
    name: DeenMate API Support
    email: api-support@deenmate.com
    url: https://docs.deenmate.com
  license:
...
servers:
...
paths:
...
  /quran/chapters/{chapterNumber}/verses:
    get:
      tags:
        - Quran
      summary: Get Chapter Verses
      description: Retrieve verses from a specific chapter with pagination. Supports upstream-compatible translation parameters.
      parameters:
        - name: chapterNumber
          in: path
          required: true
          description: Chapter ID (1-114)
          schema:
            type: integer
            minimum: 1
            maximum: 114
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Verses per page
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: language
          in: query
          description: Optional language code for upstream compatibility (e.g., en)
          schema:
            type: string
        - name: translations
          in: query
          description: Comma-separated translation resource IDs (e.g., 20,131)
          schema:
            type: string
      responses:
        '200':
          description: Chapter verses with pagination
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuranVersesResponse'
        '400':
          description: Invalid chapter ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Chapter not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /quran/translations:
    get:
      tags:
        - Quran
      summary: Get Translations
      description: Retrieve available Quran translations (optionally filter by language)
      parameters:
        - name: language
          in: query
          description: Language code to filter (e.g., en, bn)
          schema:
            type: string
      responses:
        '200':
          description: List of available translations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationsResponse'
  /quran/tafsirs:
    get:
      tags:
        - Quran
      summary: Get Tafsir Resources
      description: Retrieve available tafsir resources (optionally filter by language)
      parameters:
        - name: language
          in: query
          description: Language code to filter (e.g., en, bn)
          schema:
            type: string
      responses:
        '200':
          description: List of tafsir resources
          content:
            application/json:
              schema:
                type: object
                properties:
                  tafsirs:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        name:
                          type: string
                        language:
                          type: string
  /quran/verses/{verseKey}/tafsir:
    get:
      tags:
        - Quran
      summary: Get Tafsir for a Verse
      description: Retrieve tafsir for a specific verse (optionally filter by tafsirId)
      parameters:
        - name: verseKey
          in: path
          required: true
          description: Verse key (e.g., 2:255)
          schema:
            type: string
        - name: tafsirId
          in: query
          description: Tafsir resource ID
          schema:
            type: integer
      responses:
        '200':
          description: Tafsir content for the verse
          content:
            application/json:
              schema:
                type: object
                properties:
                  tafsirs:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        text:
                          type: string
  /quran/reciters:
    get:
      tags:
        - Quran
      summary: Get Reciters
      description: Retrieve available Quran reciters (optionally filter by language)
      parameters:
        - name: language
          in: query
          description: Language code (e.g., en)
          schema:
            type: string
      responses:
        '200':
          description: List of available reciters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecitersResponse'
  /quran/recitations/{recitationId}/by_chapter/{chapterNumber}:
    get:
      tags:
        - Quran
      summary: Recitation Audio by Chapter
      description: Retrieve verse-by-verse audio for a chapter (upstream-compatible)
      parameters:
        - name: recitationId
          in: path
          required: true
          description: Recitation (reciter) ID
          schema:
            type: integer
        - name: chapterNumber
          in: path
          required: true
          description: Chapter ID
          schema:
            type: integer
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
        - name: per_page
          in: query
          description: Items per page
          schema:
            type: integer
      responses:
        '200':
          description: Recitation audio list
          content:
            application/json:
              schema:
                type: object
                properties:
                  audio_files:
                    type: array
                    items:
                      type: object
                      properties:
                        verse_key:
                          type: string
                        url:
                          type: string
                          format: uri
  /quran/recitations/{recitationId}/by_ayah/{verseKey}:
    get:
      tags:
        - Quran
      summary: Recitation Audio by Ayah
      description: Retrieve audio for a specific ayah (upstream-compatible)
      parameters:
        - name: recitationId
          in: path
          required: true
          description: Recitation (reciter) ID
          schema:
            type: integer
        - name: verseKey
          in: path
          required: true
          description: Verse key (e.g., 2:255)
          schema:
            type: string
      responses:
        '200':
          description: Recitation audio for ayah
          content:
            application/json:
              schema:
                type: object
                properties:
                  audio_file:
                    type: object
                    properties:
                      verse_key:
                        type: string
                      url:
                        type: string
                        format: uri
  /quran/verses/{verseKey}:
    get:
      tags:
        - Quran
      summary: Get Verse
      description: Retrieve a specific verse. Supports upstream-compatible translation parameters.
      parameters:
        - name: verseKey
          in: path
          required: true
          description: Verse key (e.g., 2:255)
          schema:
            type: string
        - name: language
          in: query
          description: Optional language code for upstream compatibility (e.g., en)
          schema:
            type: string
        - name: translations
          in: query
          description: Comma-separated translation resource IDs (e.g., 20,131)
          schema:
            type: string
      responses:
        '200':
          description: Verse content
          content:
            application/json:
              schema:
                type: object
                properties:
                  verse:
                    $ref: '#/components/schemas/QuranVerse'
...
components:
  schemas:
...
    BaseResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Whether the request was successful
        data:
          description: Response data
        meta:
          description: Metadata about the response
      required:
        - success
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            success:
              const: false
            error:
              type: string
              description: Error message
            code:
              type: string
              description: Error code
            details:
              description: Additional error details
          required:
            - error
    PaginationMeta:
      type: object
      properties:
        currentPage:
          type: integer
          description: Current page number
        perPage:
          type: integer
          description: Items per page
        totalPages:
          type: integer
          description: Total number of pages
        totalItems:
          type: integer
          description: Total number of items

    # Health Schemas
...
    QuranVerse:
      type: object
      properties:
        id:
          type: integer
          description: Unique verse ID
        verseNumber:
          type: integer
          description: Verse number within the chapter
        textUthmani:
          type: string
          description: Verse text in Uthmani script
        textIndopak:
          type: string
          description: Verse text in Indopak script
        juz:
          type: integer
          description: Juz number
        hizb:
          type: integer
          description: Hizb number
        page:
          type: integer
          description: Page number in the Mushaf
      required:
        - id
        - verseNumber
        - textUthmani
...
    QuranVersesResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            success:
              const: true
            data:
              type: object
              properties:
                verses:
                  type: array
                  items:
                    $ref: '#/components/schemas/QuranVerse'
                pagination:
                  $ref: '#/components/schemas/PaginationMeta'

    # Prayer Schemas
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated endpoints

# Global security (for future use)
# security:
#   - BearerAuth: []